+++ src/atoms/InputContainer/index.tsx
@@ -1,8 +1,18 @@
 import styles from './InputContainer.module.css';
 
-export default function InputContainer({ children, state }) {
+export type State = 'loading' | 'validated' | 'error' | 'warning';
+
+export default function InputContainer({
+	children,
+	state,
+	className
+} : {
+	children: React.ReactNode,
+	state?: State,
+	className?: string,
+}) {
 	return (
-		<div className={styles.container + (state ? ' ' + styles[state] : '')}>
+		<div className={`${styles.container}${state ? ` ${styles[state]}` : ''}${className ? ` ${className}` : ''}`}>
 			{children}
 		</div>
 	);