+++ src/molecules/UserProfileCard/index.tsx
@@ -3,7 +3,7 @@ import styles from './UserProfileCard.module.css';
 import UserAvatar from '../../atoms/UserAvatar';
 
 import { FragmentType, useFragment } from '../../types/graphql/fragment-masking';
-import { graphql } from '../../types/graphql/gql';
+import { graphql } from '../../types/graphql';
 
 const UserProfileCardFragment = graphql(`#graphql
 	fragment UserProfileCardFragment on User {